Solution by Mathe Eliel
In my solution, I read each character in the entered string, creating a new substring which is compared to words in my dictionnary. If a word matches, the output string is concantenated to substring and this substring is set to empty. We loop until the end of the input string.